Improve splash screen and link it to main
