+++ lib/main.dart
@@ -1,90 +1,24 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:cerebro_flutter/screens/splash.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(MyApp());
 
 final dummySnapshot = [
- {"name": "TechHunt", "fee": 150},
- {"name": "CS:Go", "fee": 150},
- {"name": "Big Tech Quiz", "fee": 100},
- {"name": "Flutter Hackathon", "fee": 50},
- {"name": "Typeracer", "fee": 70},
+  {"name": "TechHunt", "fee": 150},
+  {"name": "CS:Go", "fee": 150},
+  {"name": "Big Tech Quiz", "fee": 100},
+  {"name": "Flutter Hackathon", "fee": 50},
+  {"name": "Typeracer", "fee": 70},
 ];
 
 class MyApp extends StatelessWidget {
- @override
- Widget build(BuildContext context) {
-   return MaterialApp(
-     debugShowCheckedModeBanner: false,
-     title: 'Events',
-     home: MyHomePage(),
-     theme: ThemeData.dark(),
-   );
- }
-}
-
-class MyHomePage extends StatefulWidget {
- @override
- _MyHomePageState createState() {
-   return _MyHomePageState();
- }
-}
-
-class _MyHomePageState extends State<MyHomePage> {
- @override
- Widget build(BuildContext context) {
-   return Scaffold(
-     appBar: AppBar(title: Text('Event List')),
-     body: _buildBody(context),
-   );
- }
-
- Widget _buildBody(BuildContext context) {
-   return _buildList(context, dummySnapshot);
- }
-
- Widget _buildList(BuildContext context, List<Map> snapshot) {
-   return ListView(
-     padding: const EdgeInsets.only(top: 20.0),
-     children: snapshot.map((data) => _buildListItem(context, data)).toList(),
-   );
- }
-
- Widget _buildListItem(BuildContext context, Map data) {
-   final record = Record.fromMap(data);
-
-   return Padding(
-     key: ValueKey(record.name),
-     padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-     child: Container(
-       decoration: BoxDecoration(
-         border: Border.all(color: Colors.grey),
-         borderRadius: BorderRadius.circular(5.0),
-       ),
-       child: ListTile(
-         title: Text(record.name),
-         trailing: Text(record.fee.toString()),
-         onTap: () => print(record),
-       ),
-     ),
-   );
- }
-}
-
-class Record {
- final String name;
- final int fee;
- final DocumentReference reference;
-
- Record.fromMap(Map<String, dynamic> map, {this.reference})
-     : assert(map['name'] != null),
-       assert(map['fee'] != null),
-       name = map['name'],
-       fee = map['fee'];
-
- Record.fromSnapshot(DocumentSnapshot snapshot)
-     : this.fromMap(snapshot.data, reference: snapshot.reference);
-
- @override
- String toString() => "Record<$name:$fee>";
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Events',
+      home: Splash(),
+      theme: ThemeData.dark(),
+    );
+  }
 }
