+++ lib/screens/splash.dart
@@ -1,20 +1,30 @@
+import 'package:cerebro_flutter/screens/events.dart';
 import 'package:flutter/material.dart';
+import 'package:splashscreen/splashscreen.dart';
 
-class Splash extends StatelessWidget {
- @override
- Widget build(BuildContext context) {
-   return Center(
-     child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: <Widget>[
-          Image.asset('assets/images/icon.png'),
-          Image(
-            image: AssetImage('assets/images/fest-logo.png'),
-            height: 80 ,
-          )
-        ],
-     )
-   );
- }
+class Splash extends StatefulWidget {
+  @override
+  _SplashState createState() => new _SplashState();
+}
+
+class _SplashState extends State<Splash> {
+  @override
+  Widget build(BuildContext context) {
+    return new SplashScreen(
+        seconds: 7,
+        navigateAfterSeconds: EventsPage(),
+        title: Text(
+          'Just wait a little longer...',
+          style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20.0),
+        ),
+        image: Image(
+          image: AssetImage('assets/images/fest-logo.png'),
+          height: 120,
+        ),
+        backgroundColor: Colors.black,
+        styleTextUnderTheLoader: TextStyle(),
+        photoSize: 100.0,
+        onClick: () => print("Clicked Logo"),
+        loaderColor: Colors.orange);
+  }
 }
